refactor(engine): move useSyntheticShadow to env/dom
